Update on "[1/n] Adding Engine, Event, EventHandler for Composable DDP"
Step 1 for https://github.com/pytorch/pytorch/issues/37002

This commit adds minimum e2e testable Engine, Event, and
EventHandler implementations. A majority of these implementations
are WIP, but should be sufficient to allow subsequent works to
be implemented concurrently.

This commit implements a bare-minimum Engine that builds and
verifies the event-handling graph, and routes events to consumer
EventHandler instances. It also implements WIP default
EventHandlers, e.g., DefaultTrigger, DefaultBucketer, and
AllReduceComm.

Differential Revision: [D27605300](https://our.internmc.facebook.com/intern/diff/D27605300)

[ghstack-poisoned]
