+++ tools/build_variables.bzl
@@ -328,7 +328,6 @@ libtorch_distributed_sources = [
     "torch/csrc/distributed/rpc/utils.cpp",
     "torch/csrc/distributed/rpc/metrics/registry.cpp",
     "torch/csrc/distributed/spmd/engine.cpp",
-    "torch/csrc/distributed/spmd/event_handler_impl.cpp",
     "torch/csrc/distributed/spmd/event_impl.cpp",
     "torch/csrc/distributed/spmd/event.cpp",
 ]
@@ -640,6 +639,7 @@ libtorch_python_distributed_sources = libtorch_python_distributed_core_sources +
     "torch/csrc/distributed/rpc/testing/init.cpp",
     "torch/csrc/distributed/rpc/unpickled_python_call.cpp",
     "torch/csrc/distributed/rpc/unpickled_python_remote_call.cpp",
+    "torch/csrc/distributed/spmd/event_handler_impl.cpp",
     "torch/csrc/distributed/spmd/init.cpp",
     "torch/csrc/jit/runtime/register_distributed_ops.cpp",
 ]
